Call it a day.

+++ nn/mlp.py
@@ -8,6 +8,7 @@ from nn.utils import get_initializer
 
 logger = logging.getLogger(__name__)
 
+
 class MLP(Module):
     def __init__(self, units_list, out_size=None, layer_type='dense', 
                 norm=None, activation=None, kernel_initializer='glorot_uniform', 
@@ -33,7 +34,7 @@ class MLP(Module):
             kernel_initializer = get_initializer(kernel_initializer, gain=out_gain)
             self._layers.append(layer_cls(
                 out_size, kernel_initializer=kernel_initializer, 
-                name=f'{name}/out', **kwargs))
+                dtype=out_dtype, name=f'{name}/out', **kwargs))
 
     def reset(self):
         for l in self._layers:
